+++ frontend/app/app/features/articles/components/ArticleDetailPage/Article/Body/index.tsx
@@ -3,10 +3,22 @@
 import { CardBody, Heading, Text } from "@chakra-ui/react";
 
 import { useArticleData } from "@/app/features/articles/hooks/useArticleData";
+import { useEditMode } from "@/app/features/articles/hooks/useEditMode";
 import LinkCardWithMetaData from "@/app/features/link_preview/components/LinkCard";
+
+import UpdateArticleForm from "../../../UpdateArticleForm";
 
 const ArticleBody: React.FC = () => {
   const { title, content, portfolioSiteUrl, repositoryUrl } = useArticleData();
+  const { isEditMode } = useEditMode();
+  if (isEditMode) {
+    return (
+      <CardBody>
+        <UpdateArticleForm />
+      </CardBody>
+    );
+  }
+
   return (
     <CardBody>
       <Heading fontSize="lg" mb="3">
